Feature: Supporting the ability to generate result files with a specific fixed filename.
