+++ base/__init__.py
@@ -6,3 +6,4 @@ from .base_split_data_loader import *
 from .base_metric import *
 from .base_confusion import *
 from .base_hook import *
+from .base_resultvisualization import *
